+++ lib/src/widgets/widgets.dart
@@ -1,1 +1,2 @@
 export 'gap.dart';
+export 'flexbox.dart';